requirements.txt
astroid==1.4.8
backports-abc==0.5
backports.functools-lru-cache==1.3
certifi==2017.1.23
colorama==0.3.7
configparser==3.5.0
dj-database-url==0.4.1
Django==1.10.5
django-cors-headers==2.0.1
django-filter==1.0.1
django-password-reset==0.9
django-registration==2.2
django-tinymce==2.4.0
djangorestframework==3.5.3
gunicorn==19.6.0
httpie==0.9.9
isort==4.2.5
lazy-object-proxy==1.2.2
Markdown==2.6.7
mccabe==0.5.2
Pillow==3.4.2
psutil==3.4.2
Pygments==2.2.0
pylint==1.6.4
pytz==2016.7
redis==2.10.5
requests==2.13.0
singledispatch==3.4.0.3
six==1.10.0
sqlparse==0.1.18
tornado==4.4.2
virtualenv==15.1.0
virtualenvwrapper-win==1.2.1
whitenoise==3.2.2
wrapt==1.10.8
